% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRepresentativeDocs.STS.R
\name{findRepresentativeDocs}
\alias{findRepresentativeDocs}
\title{A Generic Function to Identify Documents that Load Heavily on a Topic}
\usage{
findRepresentativeDocs(object, corpus_text, topic, n = 3)
}
\arguments{
\item{object}{Model output from sts}

\item{corpus_text}{vector of text documents, usually contained in the output of prepDocuments}

\item{topic}{a single topic number}

\item{n}{number of documents to extract}
}
\description{
This is a generic function that extracts documents with the
highest prevalence for a given topic.
}
\seealso{
\code{\link{findRepresentativeDocs.STS}} for the method.
}
