% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTopWords.STS.R
\name{printTopWords}
\alias{printTopWords}
\title{A Generic Function to Print Top Words}
\usage{
printTopWords(object, n = 10, lowerPercentile = 0.05, upperPercentile = 0.95)
}
\arguments{
\item{object}{Model output from sts}

\item{n}{number of words to print to console for each topic}

\item{lowerPercentile}{Percentile to calculate a representative negative sentiment document.}

\item{upperPercentile}{Percentile to calculate a representative positive sentiment document.}
}
\description{
This is a generic function that prints the top words for each
document for low, average, and high levels of
sentiment-discourse
}
\seealso{
\code{\link{printTopWords.STS}} for the method.
}
