% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.STS.R
\name{summary.STS}
\alias{summary.STS}
\alias{print.STS}
\title{Summary Function for the STS objects}
\usage{
\method{summary}{STS}(object, ...)
}
\arguments{
\item{object}{An STS object.}

\item{\dots}{Additional arguments affecting the summary}
}
\description{
Function to report on the contents of STS objects
}
\details{
Summary prints a short statement about the model and then runs
\code{\link{printTopWords}}.
}
